+++ src/app/feat-utils/voice/voice_locker.ts
@@ -1,19 +1,31 @@
 import { setTimeout } from 'timers/promises';
 
-import { EmbedBuilder, ActionRowBuilder, ButtonBuilder, ButtonStyle } from 'discord.js';
+import {
+    EmbedBuilder,
+    ActionRowBuilder,
+    ButtonBuilder,
+    ButtonStyle,
+    ButtonInteraction,
+    ChatInputCommandInteraction,
+    VoiceState,
+    Channel,
+} from 'discord.js';
 
 import { log4js_obj } from '../../../log4js_settings';
-import { exists, notExists } from '../../common/others';
+import { getGuildByInteraction } from '../../common/manager/guild_manager';
+import { searchAPIMemberById } from '../../common/manager/member_manager';
+import { assertExistCheck, exists, notExists } from '../../common/others';
 const logger = log4js_obj.getLogger('interaction');
 
 /*
  * スラコマ打たれたときの動作
  */
-export async function voiceLocker(interaction: $TSFixMe) {
-    if (!interaction.inGuild()) return;
-
-    const author = interaction.member;
+export async function voiceLocker(interaction: ChatInputCommandInteraction<'cached' | 'raw'>) {
+    const guild = await getGuildByInteraction(interaction);
+    const author = await searchAPIMemberById(guild, interaction.member.user.id);
+    assertExistCheck(author, 'author');
     const channel = interaction.channel;
+    assertExistCheck(channel, 'channel');
     const limitNum = interaction.options.getInteger('人数');
 
     // ボイスチャンネル未接続or違うボイスチャンネル接続中だと弾く
@@ -43,36 +55,56 @@ export async function voiceLocker(interaction: $TSFixMe) {
             isLock: limitNum == 0 ? false : true,
         };
 
-        // 制限人数を反映
-        channel.setUserLimit(limitNum);
+        if (channel.isVoiceBased()) {
+            // 制限人数を反映
+            await channel.setUserLimit(limitNum);
+        }
     } else {
-        channelState = await getVoiceChannelState(interaction);
+        channelState = await getVoiceChannelState(channel);
     }
 
-    const embed = createEmbed(channelState);
-    const button = createButton(channelState);
+    if (exists(channelState)) {
+        const embed = createEmbed(channelState);
+        const button = createButton(channelState);
 
-    await interaction
-        .reply({
-            embeds: [embed],
-            components: [button],
-            fetchReply: true,
-        })
-        .catch((error: $TSFixMe) => {
-            logger.error(error);
+        await interaction
+            .reply({
+                embeds: [embed],
+                components: [button],
+                fetchReply: true,
+            })
+            .catch((error) => {
+                logger.error(error);
+            });
+
+        // 1分後にメッセージを削除
+        await setTimeout(60000);
+        await interaction.deleteReply();
+    } else {
+        await interaction.reply({
+            content: '`[ERROR]`チャンネルが見つからなかったでし！',
         });
-
-    // 1分後にメッセージを削除
-    await setTimeout(60000);
-    await interaction.deleteReply();
+    }
 }
 
 /*
  * ボタンが押されたときの動作
  */
-export async function voiceLockerUpdate(interaction: $TSFixMe) {
-    const member = interaction.member;
+export async function voiceLockerUpdate(interaction: ButtonInteraction<'cached' | 'raw'>) {
+    const guild = await getGuildByInteraction(interaction);
+
+    const member = await searchAPIMemberById(guild, interaction.member.user.id);
+    assertExistCheck(member, 'member');
     const channel = interaction.channel;
+    assertExistCheck(channel, 'channel');
+
+    if (!channel.isVoiceBased()) {
+        return await interaction.reply({
+            content: 'ボイスチャンネルでないと操作できないでし！',
+            ephemeral: true,
+        });
+    }
+
     // ボイスチャンネル内のメンバー数
     const voiceMemberNum = channel.members.size;
 
@@ -85,19 +117,21 @@ export async function voiceLockerUpdate(interaction: $TSFixMe) {
         return;
     }
 
-    const channelState = await getVoiceChannelState(interaction);
+    const channelState = await getVoiceChannelState(channel);
 
-    let limit = Number(channelState.limit);
+    if (notExists(channelState)) return;
+
+    let limit = channelState.limit;
 
     // 'LOCK'ボタンor'UNLOCK'ボタンを押したとき
     if (interaction.customId == 'voiceLockOrUnlock') {
         const label = interaction.component.label; // ボタンのラベルから設定する状態を取得
         if (label === 'LOCK') {
-            await interaction.channel.setUserLimit(voiceMemberNum);
+            await channel.setUserLimit(voiceMemberNum);
             channelState.isLock = true;
             channelState.limit = voiceMemberNum;
         } else if (label === 'UNLOCK') {
-            await interaction.channel.setUserLimit(0);
+            await channel.setUserLimit(0);
             channelState.isLock = false;
             channelState.limit = 0;
         }
@@ -110,14 +144,14 @@ export async function voiceLockerUpdate(interaction: $TSFixMe) {
             if (limit != 99) {
                 limit += 1;
                 channelState.limit = limit;
-                await interaction.channel.setUserLimit(limit);
+                await channel.setUserLimit(limit);
             }
         } else if (interaction.customId === 'voiceLock_dec') {
             // 1人で押されたときは何もしない
             if (limit != 1) {
                 limit -= 1;
                 channelState.limit = limit;
-                await interaction.channel.setUserLimit(limit);
+                await channel.setUserLimit(limit);
             }
         }
     } else {
@@ -129,7 +163,7 @@ export async function voiceLockerUpdate(interaction: $TSFixMe) {
                     ephemeral: true,
                     fetchReply: true,
                 })
-                .catch((error: $TSFixMe) => {
+                .catch((error) => {
                     logger.error(error);
                 });
             return;
@@ -142,12 +176,14 @@ export async function voiceLockerUpdate(interaction: $TSFixMe) {
             components: [createButton(channelState)],
             fetchReply: true,
         })
-        .catch((error: $TSFixMe) => {
+        .catch((error) => {
             logger.error(error);
         });
 }
 
-export async function disableLimit(oldState: $TSFixMe) {
+export async function disableLimit(oldState: VoiceState) {
+    const oldChannel = oldState.channel;
+    if (notExists(oldChannel)) return;
     const usable_channel = [
         'alfa',
         'bravo',
@@ -163,32 +199,40 @@ export async function disableLimit(oldState: $TSFixMe) {
         'lima',
         'mike',
     ];
-    const oldChannel = await oldState.guild.channels.fetch(oldState.channelId);
+
     // 使用可能VCかチェック
     if (!usable_channel.includes(oldChannel.name)) {
         return;
     }
 
     if (oldChannel.members.size == 0) {
-        oldChannel.setUserLimit(0);
+        await oldChannel.setUserLimit(0);
     }
 }
 
+type ChannelState = {
+    id: string;
+    limit: number;
+    isLock: boolean;
+};
+
 /**
- * インタラクションからチャンネル情報を取得する用のオブジェクトを作成する
- * @param {*} interaction
+ * チャンネル情報のオブジェクトを作成する
+ * @param {*} channel チャンネル情報
  * @returns channelStateのオブジェクトを返す
  */
-async function getVoiceChannelState(interaction: $TSFixMe) {
-    const channel = interaction.member.voice.channel;
+async function getVoiceChannelState(channel: Channel) {
+    if (channel.isVoiceBased()) {
+        const channelStateObj: ChannelState = {
+            id: channel.id,
+            limit: channel.userLimit,
+            isLock: channel.userLimit == 0 ? false : true,
+        };
 
-    const channelStateObj = {
-        id: channel.id,
-        limit: channel.userLimit,
-        isLock: channel.userLimit == 0 ? false : true,
-    };
-
-    return channelStateObj;
+        return channelStateObj;
+    } else {
+        return null;
+    }
 }
 
 /**
@@ -196,8 +240,8 @@ async function getVoiceChannelState(interaction: $TSFixMe) {
  * @param {*} channelState チャンネル情報の読み込み
  * @returns 作成したボタンを返す
  */
-function createButton(channelState: $TSFixMe) {
-    const button = new ActionRowBuilder();
+function createButton(channelState: ChannelState) {
+    const button = new ActionRowBuilder<ButtonBuilder>();
     const limit = channelState.limit;
     if (channelState.isLock) {
         // 制限人数が1のとき，'－'ボタンを無効化
@@ -240,7 +284,7 @@ function createButton(channelState: $TSFixMe) {
  * @param {*} channelState チャンネル情報の読み込み
  * @returns 作成したEmbedを返す
  */
-function createEmbed(channelState: $TSFixMe) {
+function createEmbed(channelState: ChannelState) {
     let limit;
     // 制限人数表示用の判定
     if (channelState.limit === 0) {
